requirements.txt
supervision==0.16.0
rfdetr
redis==5.0.1
boto3==1.34.0
requests==2.31.0
Pillow==10.1.0
torch
torchvision